requirements_spaces.txt
torch>=2.1.0
diffusers==0.24.0
transformers==4.36.2
accelerate==0.25.0
gradio==4.19.0
safetensors>=0.4.0
Pillow>=10.0.0
numpy>=1.24.0
xformers>=0.0.23
huggingface_hub>=0.20.0
